refactoring & standardization: Moved graph building and executing routines by Jhoskins to _dask_graph_tools.py
